--- src/weightsys.rs
@@ -1,4 +1,0 @@
-pub struct WeightSys {
-    gravity: f64,
-    pull: (f64, f64),
-}
